refactor: move CollectiveStatsCard and HomepageSponsorCard to home folder
