+++ pages/index.js
@@ -23,9 +23,9 @@ import ListItem from '../components/ListItem';
 import Hide from '../components/Hide';
 import Container from '../components/Container';
 import StyledLink from '../components/StyledLink';
-import CollectiveStatsCard from '../components/CollectiveStatsCard';
+import CollectiveStatsCard from '../components/home/CollectiveStatsCard';
 import NewsletterContainer from '../components/NewsletterContainer';
-import HomepageSponsorCard from '../components/HomepageSponsorCard';
+import HomepageSponsorCard from '../components/home/HomepageSponsorCard';
 import Carousel from '../components/Carousel';
 import Currency from '../components/Currency';
 import ErrorPage from '../components/ErrorPage';
